Notepad profile update
Reworked the notepad so it can change as the game is played.
Notepad is its own object separate from the pause menu and is opened with tab.
Profiles have an initial set of visible objects and a final set of visible objects.
A profiles state can be changed through public methods that are accessible with unity events.