Reworked the building system
Made it so all of the logic and information is kept in a prefab.
Removed scriptable objects and separate prefabs for the area to build on and the building.
Made it so the player can destroy a building or return objects from partially made building.
Added lock so you need to unlock the ability to build.
Reworked signs to include images.
Changed items from enums to strings to make it easier.
Fixed existing prefabs.
Added prefabs back into game scene.
Removed requirement for an item from interactables.