Finishing tasks from last playtest
Made it so items need to be unlocked to be picked up. The have an item identifier which the player must have unlocked.
Add prompt in dialogue to tell the player to press E to continue.
Add prompt top tell the player when the notepad has been updated.
Add icon above NPC when they have an active task.
Cleaned up getting onto ladders so it is smoother.
Fixed the issue with interactable items highlights remaining after the player leaves the interacting distance.
Smoothed NPC rotating to talk to player.