Fixed some minor issues from playtesting
NPC looks away from player.
Removed interact animation when picking up items.
Fixed the mesh on the wood so it is on the object.
Fixed the colliders on pick up objects.
Fixed issue with player continuing jog animation while talking to NPC or climbing ladder.
Fixed issue with player falling animation running when stepping over small ledges.
Fixed controls menu so it only has controls in the game.
Added options to pause menu.
Increased the vertical clamp on the camera.