Ladders are now climbed by walking into them.
You now walk into a ladder to start climbing it.
Ladders have a bottom and top child object for the positions that you start climbing from.
Ladders have a trigger and a collider. The trigger is to detect the player and start them climbing.
Climbing animation only plays when moving.