Improve notepad feedback
Notepad now comes on to the screen from the left and then hides back off to the left.
When the notepad is updated it peeks in from the left of the screen to show it has been updated.
The icon appearing to let the player know it has been updated has been removed.