Fixes to Tom Scene
Fixed Laya profile being too big.
Fixed issue with scaffolds when unbuild having colliders that the player could get stuck in.
Fixed the issue with scaffolds particles not being on the objects.
Fixed an issue with the gate at the top of the scaffold in the electrical area being too close and the player going through it.